Fix incorrect merge conflicts
