+++ src/platform/packages/shared/kbn-monaco/src/console/index.ts
@@ -7,6 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
+import { type ESQLCallbacks, suggest } from '@kbn/esql-validation-autocomplete';
+import { MutableRefObject } from 'react';
 /**
  * This import registers the Console monaco language contribution
  */
@@ -21,6 +23,9 @@ import {
   consoleOutputLanguageConfiguration,
 } from './lexer_rules';
 import { foldingRangeProvider } from './folding_range_provider';
+import { ESQL_AUTOCOMPLETE_TRIGGER_CHARS } from '../esql';
+import { wrapAsMonacoSuggestions } from '../esql/lib/converters/suggestions';
+import { isInsideTripleQuotes } from './utils';
 
 export { CONSOLE_LANG_ID, CONSOLE_OUTPUT_LANG_ID } from './constants';
 /**
@@ -28,11 +33,50 @@ export { CONSOLE_LANG_ID, CONSOLE_OUTPUT_LANG_ID } from './constants';
  */
 export { CONSOLE_THEME_ID } from './language';
 
+export const CONSOLE_TRIGGER_CHARS = ['/', '.', '_', ',', '?', '=', '&', '"'];
+
 export const ConsoleLang: LangModuleType = {
   ID: CONSOLE_LANG_ID,
   lexerRules,
   languageConfiguration,
   foldingRangeProvider,
+  getSuggestionProvider: (
+    esqlCallbacks: Pick<ESQLCallbacks, 'getSources' | 'getPolicies'>,
+    actionsProvider: MutableRefObject<any>
+  ): monaco.languages.CompletionItemProvider => {
+    return {
+      // force suggestions when these characters are used
+      triggerCharacters: [...CONSOLE_TRIGGER_CHARS, ...ESQL_AUTOCOMPLETE_TRIGGER_CHARS],
+      provideCompletionItems: async (
+        model: monaco.editor.ITextModel,
+        position: monaco.Position,
+        context: monaco.languages.CompletionContext
+      ) => {
+        const fullText = model.getValue();
+        const cursorOffset = model.getOffsetAt(position);
+        const textBeforeCursor = fullText.slice(0, cursorOffset);
+        const { insideQuery } = isInsideTripleQuotes(textBeforeCursor);
+        if (esqlCallbacks && insideQuery) {
+          const queryStartOffset = textBeforeCursor.lastIndexOf('"""') + 3;
+          const queryText = textBeforeCursor.slice(queryStartOffset, cursorOffset);
+          const esqlSuggestions = await suggest(
+            queryText,
+            cursorOffset - queryStartOffset,
+            context,
+            esqlCallbacks
+          );
+          return {
+            suggestions: wrapAsMonacoSuggestions(esqlSuggestions, queryText, false),
+          };
+        } else if (actionsProvider.current) {
+          return actionsProvider.current?.provideCompletionItems(model, position, context);
+        }
+        return {
+          suggestions: [],
+        };
+      },
+    };
+  },
 };
 
 export const ConsoleOutputLang: LangModuleType = {
